refactor(bookmark): move getUserBookmarks and isBookmarked from BookmarkService to BookmarkQueryService; update controller usage
