+++ bookmark/src/main/java/pt/estga/bookmark/BookmarkService.java
@@ -11,8 +11,6 @@ import pt.estga.monument.MonumentRepository;
 import pt.estga.user.entities.User;
 import pt.estga.user.repositories.UserRepository;
 
-import java.util.List;
-
 @Service
 @RequiredArgsConstructor
 public class BookmarkService {
@@ -64,39 +62,6 @@ public class BookmarkService {
 
         BookmarkDto dto = mapper.toDto(bookmark);
         return new BookmarkDto(dto.id(), dto.type(), dto.targetId(), content);
-    }
-
-    public List<BookmarkDto> getUserBookmarks(Long userId) {
-        return bookmarkRepository.findAllByUserId(userId)
-                .stream()
-                .map(b -> {
-                    Object content = switch (b.getTargetType()) {
-                        case MONUMENT -> {
-                            Long parsedId = Long.parseLong(b.getTargetId());
-                            yield monumentRepository.findById(parsedId)
-                                    .map(monumentMapper::toResponseDto)
-                                    .orElse(null);
-                        }
-
-                        case MARK -> {
-                            Long parsedId = Long.parseLong(b.getTargetId());
-                            yield markRepository.findById(parsedId)
-                                    .map(markMapper::toDto)
-                                    .orElse(null);
-                        }
-
-                        default -> null;
-                    };
-                    BookmarkDto dto = mapper.toDto(b);
-                    return new BookmarkDto(dto.id(), dto.type(), dto.targetId(), content);
-                })
-                .toList();
-    }
-
-    public boolean isBookmarked(Long userId, TargetType type, String targetId) {
-        return bookmarkRepository
-                .findByUserIdAndTargetTypeAndTargetId(userId, type, targetId)
-                .isPresent();
     }
 
     @Transactional
